refactor: test에 사용되는 RepositoryTest 클래스를 하나로 통합한다. (추후 리팩터링을 위함)
